requirements.txt
# graphene==3.2.1
# graphene-sqlalchemy==3.0.0b1
# graphql-core==3.2.3
# graphql-relay==3.2.0
# graphql-server==3.0.0b5
# SQLAlchemy==1.3.24

Flask
Flask_SQLAlchemy
Flask-GraphQL
graphene>=2.0
graphene-sqlalchemy
psycopg2-binary